requirements.txt
cohere
python-telegram-bot
